+++ src/renderer/containers/Validator/Validator.stories.tsx
