+++ src/renderer/containers/Validator/validator.scss
@@ -98,7 +98,7 @@
     }
 }
 
-.validator-simple-buttons{
+.validator-buttons{
     margin-top: 25px;
     display: flex;
     justify-content: flex-end;
